Assignments/02._Data/00._Data_parsing_server_Part_II/pyproject.toml
[project]
name = "00-data-parsing-server-part-ii"
version = "0.1.0"
description = ""
authors = [
    {name = "tobiasbkraemer",email = "tobi4533@stud.keadk"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "uvicorn (>=0.34.0,<0.35.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]

[tool.poetry]
packages = [{include = "00"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
